Refactor the code generation
Introduce a new BIR instruction `FunctionConstructInsn`. We still
generate the function code for lambdas in the "front" when generating the top
level functions, but in the nBack codeGen for lambdas only happen as a
part of this instruction.

This means we need to pass a few states down (mostly via the scaffold).